ContextCache can create a Context with any Integrator.
In many cases we want to spawn a Context only to compute energies inside a function, which means that we don't care about the integrator used. I made the integrator argument optional in ContextCache.get_context that looks for cached Contexts consistent to the given thermodynamic state before creating a new one.