Added optional restart for IntegratorMoves.
Moves inheriting from BaseIntegratorMove check that energy is not NaN after integration. If it is, an optional n_restart_attempt parameter in the constructor (default 0) determine how many times the move should be restart in an attempt to recover. The error raised IntegratorMoveError can be serialized by the application for debugging.