Track external protocol work in LangevinSplittingIntegrator
Why:

Some codebases, for instance protons and saltswap, modify context
parameters externally. To save expensive energy computations,
we want to accumulate the protocol work inside the integrator.
This commit adds a `protocol_work` variable to keep track of it.

It also adds a unit test to make sure the new feature works adequately.

This change addresses the need by:

* Adding the `measure_protocol_work` parameter, default false, and
including the machinery that calculates it.
* Adding a unit test to ensure it works as expected.