Implement GlobalParameterState facility.
This class will simplify a lot the implementation of AlchemicalState, RestraintState, and all other composable states that require controlling a global parameter.
This class also add the logic to control parameters that have a specific suffix appended to their name. This feature will enable multi-restraint and multi-alchemical-region simulations in the future.